Add improvements to chat-typing-indicator examples markdown descriptions
