+++ examples/chat-room-reactions/javascript/index.html
@@ -16,13 +16,19 @@
         >
         </div>
         <div class="border-t-2 border-gray-200 px-4 pt-4 mb-2 sm:mb-0">
+          <div id="emoji-selector" className="emoji-selector"></div>
+          <div className="absolute right-0 bottom-[100px] mb-2 mr-2">
+            <div className="reaction-container">
+              <div id="reaction-area" className="reaction-area"></div>
+            </div>
+          </div>
           <form
             class="flex"
           >
             <input
                type="text"
               disabled
-              placeholder="Type something..."
+              placeholder="Start typing..."
               class="w-full focus:outline-none focus:placeholder-gray-400 text-gray-600 placeholder-gray-600 pl-2 pr-2 bg-gray-200 rounded-l-md py-1 disabled:cursor-not-allowed disabled:italic"
               autoFocus
             />
@@ -44,12 +50,6 @@
               </button>
             </div>
           </form>
-          <div class="absolute bottom-0 right-0 mb-2 mr-2">
-            <div class="reaction-container">
-              <div id="reaction-area" class="reaction-area"></div>
-              <div id="emoji-selector" class="emoji-selector"></div>
-            </div>
-          </div>
        </div>
      </div>
    </div>
