+++ examples/chat-room-reactions/react/styles/styles.css
@@ -1,3 +1,7 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
 .container {
   width: 100%;
   display: flex;
@@ -6,41 +10,6 @@
   position: relative;
   background-color: #f4f8fb;
   height: 100vh;
-}
-
-.inner {
-  width: 100%;
-  max-width: 320px;
-  padding: 1rem 0.5rem;
-}
-
-/* InputCell */
-.input-cell-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 1rem;
-}
-
-.input-cell-container:last-child {
-  display: none;
-}
-
-@media (min-height: 400px) {
-  .input-cell-container:last-child {
-    display: flex;
-  }
-}
-
-.inputContainer {
-  position: relative;
-}
-
-label {
-  margin-bottom: 0.5rem;
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: black;
 }
 
 input {
@@ -58,6 +27,7 @@ input {
   display: flex;
   flex-direction: column;
   height: 400px;
+  width: 200px;
 }
 
 .reaction-area {
@@ -70,6 +40,7 @@ input {
   display: flex;
   justify-content: space-around;
   padding: 0;
+  width: 200px;
 }
 
 .reaction {
@@ -77,7 +48,7 @@ input {
   bottom: 0;
   left: 50%;
   transform: translateX(-50%);
-  animation: moveUp 5s linear, waveSideToSide 1.5s ease-in-out;
+  animation: moveUp 5s linear, waveSideToSide 2s infinite;
 }
 
 .emoji-btn {
@@ -103,15 +74,15 @@ input {
 
 @keyframes waveSideToSide {
   0%, 100% {
-    transform: translateX(-50%) rotate(0deg);
+    transform: translateX(-50%);
   }
   25% {
-    transform: translateX(-60%) rotate(-10deg);
+    transform: translateX(-80%);
   }
   50% {
-    transform: translateX(-50%) rotate(0deg);
+    transform: translateX(-50%);
   }
   75% {
-    transform: translateX(-40%) rotate(10deg);
+    transform: translateX(-20%);
   }
 }
